requirements.txt
apache-airflow==2.7.1
pandas
psycopg2-binary
SQLAlchemy
